requirements.txt
ollama>=0.4.0
openai>=1.0.0
any-guardrail==0.2.2
guardrails-ai>=0.6.0
giskard>=2.0.0
pandas>=2.0.0
ipython>=8.0.0
python-dotenv>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
notebook>=7.4.7
